py-sphinx: add variant `sphinxcontrib`
Controls whether the Sphinx contrib components are installed.

Motivation: As of this commit date, the source tarballs for these two
packages both trigger the ClamXAV 3.5.5 malware detector:
- py-sphinxcontrib-applehelp 1.0.4
- py-sphinxcontrib-htmlhelp 2.0.1

This is almost certainly a false positive.  Regardless, by declaring the
runtime dependencies within a variant, it becomes possible to install
port py-sphinx without the contrib components, and therefore without
triggering the ClamXAV malware detector, until such time as the malware
detection database is updated, or the source tarballs are updated.